Cleanups based on comments
