+++ src/markers/index.js
@@ -1,1 +1,2 @@
+export * from "./Cluster";
 export * from "./Pin";
